@@ -19,27 +19,28 @@
   "author": "Oliver Tacke",
   "license": "MIT",
   "devDependencies": {
-    "@babel/cli": "^7.23.9",
-    "@babel/core": "^7.23.9",
-    "@babel/plugin-transform-destructuring": "^7.23.3",
-    "@babel/preset-env": "^7.23.9",
+    "@babel/cli": "^7.24.1",
+    "@babel/core": "^7.24.4",
+    "@babel/plugin-transform-destructuring": "^7.24.1",
+    "@babel/preset-env": "^7.24.4",
     "babel-loader": "^9.1.3",
-    "css-loader": "^6.10.0",
-    "eslint": "^8.56.0",
-    "eslint-plugin-jsdoc": "^48.0.4",
-    "mini-css-extract-plugin": "^2.8.0",
+    "css-loader": "^7.1.0",
+    "eslint": "^9.0.0",
+    "eslint-config-snordian-h5p": "github:otacke/eslint-config-snordian-h5p",
+    "mini-css-extract-plugin": "^2.8.1",
     "regenerator-runtime": "^0.14.1",
-    "sass": "^1.70.0",
-    "sass-loader": "^14.1.0",
-    "stylelint": "^16.2.1",
+    "sass": "^1.74.1",
+    "sass-loader": "^14.1.1",
+    "stylelint": "^16.3.1",
     "stylelint-config-recommended": "^14.0.0",
     "stylelint-config-recommended-scss": "^14.0.0",
-    "webpack": "^5.90.1",
+    "webpack": "^5.91.0",
     "webpack-cli": "^5.1.4"
   },
   "browserslist": "> 0.5%, last 2 versions, Firefox ESR, not dead",
   "dependencies": {
     "color": "^4.2.3",
     "he": "^1.2.0"
-  }
+  },
+  "type": "module"
 }
