@@ -1,21 +1,26 @@
-const path = require('path');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const TerserPlugin = require('terser-webpack-plugin');
+import { dirname, resolve as _resolve, join } from 'path';
+import { fileURLToPath } from 'url';
+import MiniCssExtractPlugin from 'mini-css-extract-plugin';
+import TerserPlugin from 'terser-webpack-plugin'; // Provided by webpack
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 const mode = process.argv.includes('--mode=production') ?
-  'production' : 'development';
+  'production' :
+  'development';
 const libraryName = process.env.npm_package_name;
 
-module.exports = {
+export default {
   mode: mode,
   resolve: {
     alias: {
-      '@assets': path.resolve(__dirname, 'src/assets'),
-      '@components': path.resolve(__dirname, 'src/scripts/components'),
-      '@mixins': path.resolve(__dirname, 'src/scripts/mixins'),
-      '@models': path.resolve(__dirname, 'src/scripts/models'),
-      '@scripts': path.resolve(__dirname, 'src/scripts'),
-      '@services': path.resolve(__dirname, 'src/scripts/services')
+      '@assets': _resolve(__dirname, 'src/assets'),
+      '@components': _resolve(__dirname, 'src/scripts/components'),
+      '@mixins': _resolve(__dirname, 'src/scripts/mixins'),
+      '@models': _resolve(__dirname, 'src/scripts/models'),
+      '@scripts': _resolve(__dirname, 'src/scripts'),
+      '@services': _resolve(__dirname, 'src/scripts/services')
     }
   },
   optimization: {
@@ -40,7 +45,7 @@ module.exports = {
   },
   output: {
     filename: `${libraryName}.js`,
-    path: path.resolve(__dirname, 'dist'),
+    path: _resolve(__dirname, 'dist'),
     clean: true
   },
   target: ['browserslist'],
@@ -70,12 +75,12 @@ module.exports = {
       },
       {
         test: /\.svg|\.jpg|\.png$/,
-        include: path.join(__dirname, 'src/assets/images'),
+        include: join(__dirname, 'src/images'),
         type: 'asset/resource'
       },
       {
         test: /\.woff$/,
-        include: path.join(__dirname, 'src/assets/fonts'),
+        include: join(__dirname, 'src/fonts'),
         type: 'asset/resource'
       }
     ]
